alarms: add ndc info to alarm info
Motivation:

Sometimes information, like the pnfsid, is missing from a logging
message which we promote to alarm, but the information is present
in the MDC (NDC):

rw-stkendca13a-5Domain.log:20 Jul 2016 09:30:40 (rw-stkendca13a-5) [b55b5e9b-baa0-4613-b56c-e10c32d71d19 000013A9AA57B8684A4AA5E407F9A0F5E795] HSM script failed (script reported: 1: ...

This information can be usefully displayed in the
webadmin table for diagnostic purposes.

Modification:

Instead of appending the information on the client end (to the
event message), we allow the server-end to do so.

Result:

The missing diagnostic info is displayed in the table.

While not a bug, the improvement is minimally invasive, affects
only the displayed element, and would be useful to have.
Hence I am requesting a backport.
I leave it to the reviewers' discretion.

Target: master
Acked-by: Paul