dcache: Add framework for dynamic pool manager handlers
Motivation:

In some cases it may be possible to process certain pool manager requests
on the client side by caching a snapshot of the pool manager state in e.g.
the door. When this is possible may be dynamic and depend on the configuration
of the pool manager.

The situation is further complicated by space manager acting as a proxy
for pool manager, thus even though the pool manager's processing could
be done in the door, the space manager still needs to know about the request
- e.g. write pool selection could run in the door, except that space manager
still needs to intercept the write pool selection messages usually sent to
pool manager.

Finally, with the future introduction of redundant pool managers, there is
a need for selecting a pool manager through rendevouz hashing - in that case
we need to embed pool manager selection logic into the doors to select the
appropriate backend for the request.

All this screams for an abstraction layer for how we interact with the pool
manager backend.

Modification:

Introduces the PoolManagerHandler interface. This simple interface allows
PoolManagerMessage requests to be submitted for pool manager processing as well
as PoolIoFileMessage requests to start movers on pools. The interface has two
versions of those methods: One for asynchronous request using a
ListenableFuture and one without an explicit callback that will deliver the
result as a regular cell message to a cell. The latter has the benefit of not
requiring a callback to be created for every request; that's particularly
important for space manager, which would proxy a lot of request concurrently.
It also allows for easy forwarding of intercepted requests as it accepts the
envelope to use.

The trivial implementation of this interface simply delivers those messages
to a named pool manager; RemotePoolManagerHandler does exactly that.

To allow the specific handler used to be injected into doors dynamically, a
serializable subinterface called SerializablePoolManagerHandler adds methods
for binding the deserialized instance to a local CellStub as well as providing
a version identifier to detect when the internal configuration of a handler has
changed.

Clients query pool manager (or space manager if it proxies the request) using a
PoolManagerGetHandler request; it returns a SerializablePoolManagerHandler for
the client to use. The client queries for an updated handler using
PoolManagerGetUpdatedHandler, providing the version identifier of the existing
handler. The pool manager will not respond to this request until it got an
updated version of the handler. This allows any change to the pool manager
configuration to be instantly propagated, while allowing the client to use a
convenient callback to obtain the response. All this is encapsulated in
PoolManagerHandlerSubscriber, which also implements the PoolManagerHandler
interface to allow the client to use the subscriber as a proxy to whatever
dynamic handler was fetched remotely.

Currently, pool manager returns a RemotePoolManagerHandler and thus never
delegates any processing to the client. Space manager intercepts the request
for handler instances and decorates the handler with its own processing; the
decorator knows what type of requests the space manger needs to intercept and
other requests are passed through to the handler obtained from the real pool
manager. Thus read requests bypass space manager, and PoolAcceptFileMessage and
DoorTransferFinishedMessage for writes outside linkgroups do too.

None of the doors or other clients of pool manager have been updated yet, so
as of now doors behave just as before and do not use a pool manager handler.
This also means that the above bypass of space manager is not currently used
by any doors. This is simply to keep the patch size down and will be fixed
in a future patch.

Result:

The user visible changes of this patch should be minimal.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9560/