pool: added CEPH back-ended FileStore implementation
Motivation:
CEPH is a distributed object store with object, block and file interfaces.
It's widely used as a building block for scalable and reliable storage.

Modification:
Add CephFileStore class, which is implementing FileStore interface and
stores it's data in CEPH by using block device interface. The implementation is
based on JNI interface and depends on librados and librdb. Each dcache
pool is represented by corresponding CEPH pool. The CEPH pool must be
created in advance. For example to a pool dcache-pool-A:

$ rados mkpool dcache-pool-A

NOTICE, the pool still requires a metadata store, which as before can
be configured to use BerkeleyDB or plain files.

New properties added:

(one-of?ceph|posix)pool.backend = posix

pool.backend.ceph.cluster = admin

pool.backend.ceph.config = /etc/ceph/ceph.conf

LIMITATIONS:
As CEPH does not provides size of the pool, dCache pool size is assumed
to be 2^31 bytes.

Result:
dCache pools can use CEPH as a back end store for data files.

This is not a production-ready implementation, but good enough for testing
and as technology preview.

Acked-by: Albert Rossi
Acked-by: Gerd Behrmann
Acked-by: Paul Millar
Target: master
Require-book: yes
Require-notes: yes