poolmanager: Make pool manager replicable
Motivation:

Pool manager poses a challenge for replication as it has lots of internal
state.  A previous patch already allowed the dynamic configuration to be
synchronized over multiple instances, however stage and p2p requests are more
expensive to synchronize through ZooKeeper as they would change frequently.

Modification:

Introduces a pool manager handler that uses rendezvous hashing to select the
appropriate backend for a particular request. This ensures that as long as the
list of backends doesn't change, requests for a particular pnfs id is always
sent to the same backend. If a pool manager instance goes down or a new
instance becomes available, at most 1/n of the PnfsId space is reassigned.
While this may cause duplicate staging and replication, this property of
rendezvous hashing minimizes the impact.

A PoolManagerHandlerPublisher bean is added to pool manager to register the
instance in ZooKeeper and provide instances of the rendezvous pool manager
handler to doors. It takes care of monitoring the list of instances registered
with ZooKeeper and inform the doors of any changes.

Result:

Pool manager has been made replicable. This means one can have several pool
manager instances with a common service name. Requests are distributed over
available instances. In case of changes in the available pool managers (removal
or addition of a pool manager instance), some duplicate staging or pool to
pool replication may occur.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9574/