cells: handle empty string pool value on staging in TransferObserver
Motivation:

Changes from 2.12 to 2.13 affecting the TransferObserverV1 class introduced a
small bug.  Previously, if the pool on the request message was either "<unknown>"
(from the DCAP door) or NULL (from all other Transfers), this was interpreted,
in the absence of a mover, to mean a staging from tape.

However, this pool value can now also be an empty string.

Modification:

Strings.emptyToNull on the pool name.  (the serial id was also missing on the
bean class).

Result:

Transfers which are staging from non-DCAP doors are correctly indicated (in yellow)
on the active transfers page, instead of showing up as "No Mover found" (in red).

Target: master
Acked-by: Olufemi
Acked-by: Paul
Request: 2.16
Request: 2.15
Request: 2.14
Request: 2.13