alarms: force alarms-only-option
Motivation:

In the current alarms server configuration, there
appear two options, one for storing only alarms to
the underlying database (as opposed to all received
events), and the other for keeping a plain-text
record of all alarms (a kind of history) independent
of the current contents of the database (in the case
of XML, this would conceivably be periodically truncated
for performance reasons).

It has been our experience, however, that cluttering
the database with such events (system-wide error
logging can accumulate quite a lot of messages)
is of little practical usefulness. Exposure of the logs
through the web page is in some ways more awkward
than just grepping the log file.  The web page, which
is the primary consumer of the database, was really meant
as a checklist for alarms only.  This should also
be the purpose of the database.

Once only alarms are stored, the history file
has no further use (at least as it is currently conceived).

Modification:

The two options have been removed.  Only alarms go
to the store. The history log is now used as an
optional way of capturing all events that arrive
at the server (independently of internal server logging),
and thus is controlled by the root-level logging setting.
The option is understood as primarily for diagnostic purposes.

Result:

Less database clutter.

This patch will be followed by a modification of the
web page to expose only alarms.

Target: master
Acked-by: Tigran
Require-book: yes
Require-notes: yes