webadmin: remove error and autorefresh options from alarms page
Motivation:

Patch 9553 removed the option to store non-alarms in
the database.  This patch is the front-end counterpart.

Note that 8568 attempted to ameliorate the slow loading
of the alarm page when the default query returns many
entries (it actually neglected to do the most important
thing, add an index to the alarm attribute in the table,
though this would only affect the RDBMS implementation).
With the current change, those measures (disabling
auto refresh and setting a range limit) are no longer
necessary.

Modification:

All code relevant to the exposure of non-alarms has been
eliminated.  The changes made in 8568 have been reversed.

Result:

Only alarms are visible in the webpage table, and the
the scalability issues concerning page loading are
greatly reduced.

Target: master
Acked-by: Tigran
Require-book: yes
Require-notes: yes