srm: Include session identifier in error message when srmRm aborts an upload
Motivation:

We often see "Upload aborted because the file was deleted by another request." in
billing logs with no way for us to know who and from which host the remove request
came.

Modification:

Include the session id in the error message. The session ID itself doesn't reveal
anything, however an admin can use the ID to find the remove request in the SRM
access log.

Result:

Modified the error and billing message used when an srmRm request aborts an upload.
The message now includes the session ID of the remove request. This session ID
may be used to identify the request in the access log.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.16
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9586/