scripts: update canonicalising function
Motivation:

The dcache script contains a function that provides the canonical
absolute path from a path that may contain symbolic links; on platforms
with `readlink` command accepts the `-f` option, that is used otherwise
similar functionality is obtained.

The current fallback is not ideal as symbolic links as directories are
not followed and the code relies on parsing the output of the ls
command.

Modification:

Use the `-P` option to pwd to resolve symbolic link directories and use
`readlink` command to obtain the value of a symbolic link.

Target: master
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/9590/
Acked-by: Gerd Behrmann