cells: Prepare to change representation of unqualified cell addresses
Motivation:

An uqualified cell address is currently represented with the domain name
being "local". Besides excluding "local" to be used as a domain name, this
is also confusing as the cell may not actually be local to the current
domain.

Modification:

It would be logical to represent lack of a domain name as null, however we
have to maintain serialization compatibility for the CellAddressCore class.

This patch adds a readResolve method that ensures that if in the future
we deserialize an instance with domain being null, we internally change this
to a compatible format. After the next golden release we can change
the representation to null and reverse the role of readResolve.

Result:

No user visible changes.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/9636/