billing: Make billing text files self describing with format headers
Motivation:

Billing text files by default have a somewhat obscure format. Further, changing the
default format will most likely break existing parsers.

Modification:

This adds the configurable format strings as headers to each billing file. In the
future parsers may make use of this information to recognize the format of the
entries in the file.

A new set of headers is added upon every restart to account for the case in which
the format strings are changed.

Updated the billing indexer to ignore lines starting with a hash symbol.

The patch ports the IO code of the billing cell to Java 8 nio 2 constructs.

Result:

Billing files now include the configured format strings as headers. These headers
are included at the top of every billing file as well as in the middle in case
the billing service is restarted.

Header lines start with a #. Parsers should be updated to ignore such lines.
Format headers start with a double ##, followed by white space, the message
name, whitespace, and the actual format string.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9611/