httpd: Introduce topic for discovering services to monitor
Motivation:

httpd has a configurable list of cells to monitor. In particular when
using replicated services, this system is fragile as there may be
more than a single consumer on monitored named queue.

Modification:

Introduces a topic, WatchedServiceTopic, for httpd to discover services
to monitor.

Introduced a new dcache.topic main branch for topic names analogous to
dcache.queue. Other topics ought to eventually be migrated to this
scheme.

One could have considered alternative designs, such as watched services
broadcasting on a topic or using ZooKeeper. The justification for the
chosen design is because it is an easy evolution of what we already got.

Result:

Introduced a new topic for httpd to discover services to monitor. One
side effect is that space manager becomes monitored automatically when
discovered.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Albert Rossir <arossi@fnal.gov>

Reviewed at https://rb.dcache.org/r/9629/