webadmin: Use topic to discover services to watch
Motivation:

The old admin pages use a topic to discover services to monitor. Webadmin until
now still used a hard-coded list of services and failed to work with replicated
services.

Modification:

Allow the webapp's beans to be post processed by the hosting universal spring
cell. This allows us to remove an ugly hack that was used for webadmin to
discover doors.

Port the CellStatusCollector to multicast ping messages on the WatchedServiceTopic
to discover services to monitor. This relies on being able to receive the
replies, which necessitates the change to bean post processing.

Eventually the webapp handler may want to create a dedicated cell instead of
exposing the hosting cells bean factory. This would avoid problems with command
listener conflicts.

The patch had to refactor bits of the webadmin to ensure that it doesn't try
to query other cells until the httpd cell has started.

Result:

Webadmin now discovers services to monitor using a topic. Thus it is able to
discover space manager and replicated services. The downside is that it will
not show services unless it has been able to observe them, thus pool manager,
pnfs maanger and gplazma will not appear in the webadmin until they are online.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Albert Rossi <arossi@fnal.gov>

Reviewed at https://rb.dcache.org/r/9630/