Enable automatic detection of Postgresql master
Motivation:

When using streaming replication with Postgresql, one server is the master
while one or more servers are slaves. When the master fails, a slave can be
promoted to the new master. Until now this has required manual reconfiguration
and restart of dCache.

Modification:

Add a JDBC URL option to indicate that dCache only wants to connect to the
master (only the master allows writes). Nodes of a postgresql cluster can
be specified as a comman seperated list in the *.db.host options.

I had to obsolete alarms.db.rdbms.type as adding the targetSeverType to the
JDBC URI made the URI postgresql specific and it didn't make sense to
separate it from the type.

Result:

When used with postgresql, the *.db.host options now accept a comma separated
list of hosts. This can be used to identify all servers in a postgresql
cluster using streaming replication. dCache automatically detects the current
master. If the master changes, dCache switches over to the new master.

The option alarms.db.rdbms.type is obsolete.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9645/