cells: Fix problem with duplicate entries in source path
Motivation:

When sending a message, the cell nucleus adds the local cell as a source address
if the destination is the first in the destination path.

This logic doesn't work if the same envelope is send multiple times as it would
add the source every time.

Modification:

There are two places in which an envelope is reused. One was in CellShell and that
easy to eliminate. The second is in CellAdapter in the logic to resubmit a message
upon NoRouteToCellException. Here we cannot recreate the envelope as we would
not want to send the message with a different envelope than the one we got from
the caller.

The patch adds a third flag to the CellNucleus#sendMessage methods to control whether
a source should be added. This is almost always the case except when sending a
message from a tunnel or when resubmitting an envelope after a NoRouteToCellException.

The patch removes the need for the CellMessage#touch method; since CellMessage is
not reused (except internally by cells), there is no need for touch. Instead setting
the TTL also resets the timestamp from which the TTL is measured.

The patch keeps the low level boolean flags of CellNucleus. A future patch will
add a corresponding SendFlag. This is useful in higher level services when
forwarding messages.

Result:

Fixes a problem in which the source address would be added to the source path of
a message several times. Backport is not requested as the patch touches critical
bits of cells and the problem is not severe.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9648/