cells: Expose SendFlag to suppress addition of a source path
Motivation:

Sometimes the sender of a cell message may need the reply to be sent to another
cell or to a named queue rather than a fully qualified cell address.

Modification:

Adds a NO_SOURCE flag to SendFlag to allow the sender to suppress adding the
local cell as a source flag. When used the sender is responsible for adding the
source to the envelope.

Eventually one may consider eliminating this flag by always placing the
responsibility to add the source on the creator of the envelope (which often
will be encapsulated in a CellStub), but that would be a somewhat larger
change.

Result:

No user visible changes.

Target: trunk
Require-notes no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9649/