zookeeper: work-around slow shutdown of SessionTrackerImpl
Motivation:

When shutting down dCache, the zookeeper cell is sufficiently slow that
it requires the System cell to explicitly interrupt still running
threads.

In part, this is because SessionTrackerImpl does not react quickly when
told to shutdown.  Instead, it can take up to three seconds after
shutdown is called before its non-daemon thread finally terminates.

As a side-effect, when dCache interrupts the thread, an error is logged
with a stack-strace:

16 Aug 2016 12:09:41 (zookeeper) [] Unexpected interruption
java.lang.InterruptedException: null
	at java.lang.Object.wait(Native Method) ~[na:1.8.0_101]
	at org.apache.zookeeper.server.SessionTrackerImpl.run(SessionTrackerImpl.java:146) ~[zookeeper-3.4.6.jar:3.4.6-1569965]

Modification:

Add a work-around that notifies the main loop, alerting it that the
service is not longer running.

Result:

Faster shutdown with fewer stack-traces and errors.

Target: master
Request: 2.16
Require-notes: yes
Require-book: no
Patch: https://rb.dcache.org/r/9651/
Acked-by: Gerd Behrmann