poolmanager: Fix design flaw in PoolManagerHandler implementations
Motivation:

PoolManagerHandler specifies two methods that accept an envelope for returning
the result. The JavaDoc correctly states that the result is sent to the source
of the envelope, yet when calling these methods the caller assumed that the
handler would add the local cell as a source and the implementations indeed
added the source cell implicitly when sending the request to another cell. This
works as long as an implementation sends the request to another cell, but would
fail if the request is processed locally and returned directly.

Furthermore, when forwarding messages in space manager, the space manager would
add itself as a source, forcing the reply to go back through the same instance.
Thus if that instance dies, the reply would not be deliverable even though any
instance could have processed the reply as all relevant state is in the space
manager database.

Modification:

Clarified the JavaDoc of the PoolManagerHandler. Changed the implementations
with the NO_SOURCE SendFlag to avoid that the local cell is added as a source.
Updated PoolManagerStub to add the source address to the envelope.

Updated space manager to add it's named queue as a source when forwarding mover
creation requests. Thus replies can be processed by any space manager instance.
Other requests are forwarded without adding an extra source as the space
manager does not have to intercept the reply - thus the reply can bypass the
space manager entirely.

Result:

In installations with redundant space managers, the side effects of killing a
space manager instance have been reduced. The latency of some requests being
proxied by space maanger has been reduced.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9652/