cells: Log killer threads too when cell shutdown is slow
Motivation:

When cells do not shut down as quickly as expected, SystemCell logs the stack traces
of the remaining threads. It is however just as interesting to know what the
killer thread is up to.

Modification:

Log the stack traces of the killer threads if cell shutdown is slow.

Increased the timeout from 3 seconds to 5 seconds before giving up on regular cells.
When the timeout was introduced, most cells didn't do much during shutdown, but this
has changed and it seems reasonable we should allow more time for that. In particular
the NFS door unregisters from the port map service during shutdown and this has an
internal timeout of 2 times 2 seconds.

Result:

Added more logging in case of slow shutdown. Increased the timeout for when to give
up waiting for cells to shut down.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9662/