cells: Lowered log level of failure to deliver delivery failure notifications
Motivation:

If cells is is unable to deliver a message it sends a delivery failure to the
sender. If it is unable to deliver this notification, it logs a warning about
the discarded message.

In particular during shutdown such warnings are common as a service while it is
shutting down rejects requests that were sent to it, but the requestor is gone
too. If such failures happen at other times, they would be a sign of a service
having died or of cell communication problems, but in such cases there would be
plenty of other errors in the logs.

Modification:

Lower the log level of the message described above.

Result:

The log level of messages like

19 aug. 2016 09:09:49 (Xrootd-Gerds-MBP) [] Unable to notify [>Xrootd-Gerds-MBP@dCacheDomain] about delivery failure of message sent to [>SpaceManager@local]: No route for Xrootd-Gerds-MBP@dCacheDomain in dCacheDomain.

has been lowered to info. These messages were common when shutting down dCache.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9665/