pool: Resolved circular bean dependencies leading to bugs during shutdown
Motivation:

When shutting down a pool with running movers, the post transfer service would
be shut down before running movers were killed. This caused
RejectedExecutionExceptions to be thrown when attempting to post process files,
leaing to stack traces in the log file as well as missing billing entries and
lack of notifications to doors.

Modification:

The circular dependency is caused by how the PoolV4 bean is set as a fault
listener on transfer and post transfer beans. This patch moves the error
handling to the MoverRequestScheduler. MoverRequestScheduler and IoQueueManager
are extended with methods to register listeners, and PoolV4 adds itself as a
listener during initialiation. This breaks the circular dependency.

Result:

A bug that caused pools with running movers to not shut down cleanly has
been fixed.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9671/