namespace: add nlink as an attribute
Motivation:

It is occationally useful to know when deleting a file will trigger
garbage collection of the underlying data storage.  For files, this
happens when the file has an nlink count of 1.

Modification:

Add nlink as a file attribute.

Result:

The nlink count of a file is available when querying a file.

Target: master
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/9693/
Acked-by: Gerd Behrmann