cells: Fix several shutdown related problems
Motivation:

- Prevent duplicate kills from failing

- Prevent termination of LoginManager before the factory has shut down

- Prevent that unstarted cells are invisible and thus not cleaned
  during shutdown

- Prevent logging dead or daemon threads during shutdown

- Ensure that cell startup actually terminated before marking the
  STARTING phase as over

- Do not report startup cancellation as a bug

Modification:

Most are straightfoward fixes. The biggest change is a two step registration
of the cell nucleus in cell glue - one when the cell is created and one when
it has started. Thus an unstarted cell is visible in the ps output, and a
starting cell will be killed if the domain shuts down.

Result:

Fixes several unreleased regressions in cell shutdown.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9679/