cells: Fix regression in message ID in replies
Motivation:

We recently restructured code related to how we update the uoid in message
envelopes. This introduced a regression in which the uoid isn't reset in
replies.

Modification:

Reset the uoid in CellMessage#revertDirection.

Had to change RequestContainer slightly to prevent that it touches the envelope
after it is sent - otherwise it conflicts with how some of the unit cells simulate
message passing between components.

Result:

Fixes an unreleased regression in cell messaging.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9680/