pools: Drop legacy field in PoolCostInfo
Motivation:

We have maintained the _mover field as a sum of all mover queues for backwards
compatibility. These days it is guaranteed that the extended mover hash is
populated and that a default queue is defined.

Modification:

Drop the mover field and drop the code that tests for the pressence for a
mover hash or a default queue.

Result:

More lean cost info message.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/9685/