poolmanager: Improve stability of cost cuts
Motivation:

Pool manager employs various cost cuts to steer pool selection. It uses
those cost cuts to trigger hot spot replication, fallback to staging,
file aging, and fallback to low priority links.

Cost cuts are compared to the performance cost of a performance. The
performance cost is the average fill level of its various queues. Queues with
low limits have a higher influence on the performance cost due to this. The
rational being that a queue with a high limit probably has transfers with low
impact on pool performance.

The problem is that the actions taken as a result of hitting any of the cost
cuts influence the performance cost too. In particular hot spot replication
causes the transfer to count both at the source pool and the target pool.
Furthermore, since it is sensible to have low limits on p2p queues (the are
short-distance, high throughput transfers that will likely go faster if
processed sequentially), the relative weight of such a transfer is high in the
performance cost. Thus triggering a hot spot replication may itself push the
performance cost of the source pool higher, triggering even more hot spot
replications - this will drive the cost of the receiving pools up too, possibly
causing a perputal state of high performance cost on many pools despite a
constant client load.

Hints that this is a real world problem have been observed at sites as logs
have shown that once hot spot replication is triggered, we often end up
creating several copies. This should usually not happen as we specifically try
to avoid selecting a target pool that already got too high performance cost to
serve the file. That it happens anyway is consistent with the above analysis.

Modification:

Ignore the p2p source and destination queues as well as the staging queues
from performance cost calculation. These are the queues that typically get
filled by internal actions as a consequence of client reads.

Result:

The two p2p queues (source and destination) as well as the HSM stage queue are
no longer included in the performance cost of a pool. This is to improve the
stability and consistency of the various pool manager cost cuts, including hot
spot replication. As a consequence, existing cost cut limits may have to be
retuned as the calculated performanc cust has changed.

Deployments in which direct reads from a set of pools is forbidden (and thus
trigger p2p for any file access on the pool) will no longer load balance reads
on those pools by p2p source activity - if other queues are empty, the source
pool is selected randomly among the pools containing the file.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/9686/