alarms: only send email on first alarm occurrence
Motivation:

extensions to process alarms, also modified the handling of alarms such
that this post-processing, as well as email, occurs only on the first
instance of potentially duplicate alarms.   This behavior is currently
not enforced for the stable branches with respect to alarm emails.

Modification:

Backport the change which affects the handling of alarm emails.

Result:

Alarm email notifications are sent only on the first occurrence
of given alarm (i.e., for that alarm instance's unique ID).

Target: 2.16
Request: 2.15
Request: 2.14
Request: 2.13
Acked-by: litvinse@fnal.gov