dcache-webadmin: disable AJAX autorefresh on pages using picnet table filter library
Motivation:

We have known for a while that there is a basic incompatibility between
the way the picnet filter library handles filtering and the autorefreshing
of a webpage via the Wicket AJAX command.  The problem does not seem
solvable short of altering the picnet library.

(If we were to commit to Wicket, the proper thing to do would be to
replace all these repeaters with the data tables used on other pages,
and remove dependency on the picnet library; our continued use of
Wicket, however, being uncertain ...)

Modification:

Disable autorefresh on the pages where the tables are filtered using
picnet (JQuery).  Pages displaying plots (billing, pool queues),
active transfers and alarms are unaffected.

Result:

There is no interference from autorefresh (which makes the
filter boxes disappear and/or resets the filter).

Downside:  The services, pools, and tape transfer pages must be manually refreshed.

Target: master
Request: 2.16
Request: 2.15
Request: 2.14
Request: 2.13
Acked-by: tigran.mkrtchyan@desy.de